+++ test/test_add_points_in_pointcloud.py
@@ -4,6 +4,7 @@ from pathlib import Path
 
 import geopandas as gpd
 import pytest
+from shapely.geometry import LineString, MultiPoint, Point
 
 from pdaltools import add_points_in_pointcloud
 from pdaltools.count_occurences.count_occurences_for_attribute import (
@@ -13,16 +14,24 @@ from pdaltools.count_occurences.count_occurences_for_attribute import (
 TEST_PATH = os.path.dirname(os.path.abspath(__file__))
 TMP_PATH = os.path.join(TEST_PATH, "tmp/add_points_in_pointcloud")
 DATA_LIDAR_PATH = os.path.join(TEST_PATH, "data/decimated_laz")
-DATA_POINTS_PATH = os.path.join(TEST_PATH, "data/points_3d")
+DATA_POINTS_3D_PATH = os.path.join(TEST_PATH, "data/points_3d")
+DATA_LIGNES_PATH = os.path.join(TEST_PATH, "data/lignes_3d")
 
 INPUT_PCD = os.path.join(DATA_LIDAR_PATH, "test_semis_2023_0292_6833_LA93_IGN69.laz")
-INPUT_POINTS = os.path.join(DATA_POINTS_PATH, "Points_virtuels_0292_6833.geojson")
+INPUT_POINTS_2D = os.path.join(DATA_POINTS_3D_PATH, "Points_virtuels_2d_with_value_z_0292_6833.geojson")
+INPUT_POINTS_3D = os.path.join(DATA_POINTS_3D_PATH, "Points_virtuels_0292_6833.geojson")
+INPUT_LIGNES_2D_GEOJSON = os.path.join(DATA_LIGNES_PATH, "Lignes_2d_0292_6833.geojson")
+INPUT_LIGNES_3D_GEOJSON = os.path.join(DATA_LIGNES_PATH, "Lignes_3d_0292_6833.geojson")
+INPUT_LIGNES_SHAPE = os.path.join(DATA_LIGNES_PATH, "Lignes_3d_0292_6833.shp")
 OUTPUT_FILE = os.path.join(TMP_PATH, "test_semis_2023_0292_6833_LA93_IGN69.laz")
 
 # Cropped las tile used to test adding points that belong to the theorical tile but not to the
 # effective las file extent
 INPUT_PCD_CROPPED = os.path.join(DATA_LIDAR_PATH, "test_semis_2021_0382_6565_LA93_IGN69_cropped.laz")
-INPUT_POINTS_FOR_CROPPED_PCD = os.path.join(DATA_POINTS_PATH, "Points_virtuels_0382_6565.geojson")
+INPUT_POINTS_2D_FOR_CROPPED_PCD = os.path.join(
+    DATA_POINTS_3D_PATH, "Points_virtuels_2d_with_value_z_0382_6565.geojson"
+)
+INPUT_POINTS_3D_FOR_CROPPED_PCD = os.path.join(DATA_POINTS_3D_PATH, "Points_virtuels_0382_6565.geojson")
 OUTPUT_FILE_CROPPED_PCD = os.path.join(TMP_PATH, "test_semis_2021_0382_6565_LA93_IGN69.laz")
 
 
@@ -31,16 +40,42 @@ def setup_module(module):
 
 
 @pytest.mark.parametrize(
-    "epsg",
+    "input_file, epsg",
     [
-        "EPSG:2154",  # should work when providing an epsg value
-        None,  # Should also work with no epsg value (get from las file)
+        (INPUT_POINTS_2D, "EPSG:2154"),  # should work when providing an epsg value + GeoJSON POINTS 2D
+        (INPUT_POINTS_3D, "EPSG:2154"),  # should work when providing an epsg value + GeoJSON POINTS 3D
+        (INPUT_POINTS_2D, None),  # Should also work with no epsg value (get from las file) + GeoJSON POINTS 2D
+        (INPUT_POINTS_3D, None),  # Should also work with no epsg value (get from las file) + GeoJSON POINTS 3D
     ],
 )
-def test_clip_3d_points_to_tile(epsg):
-    # With EPSG
-    points_clipped = add_points_in_pointcloud.clip_3d_points_to_tile(INPUT_POINTS, INPUT_PCD, epsg, 1000)
+def test_clip_3d_points_to_tile(input_file, epsg):
+    points_input = gpd.read_file(input_file)
+    points_clipped = add_points_in_pointcloud.clip_3d_points_to_tile(points_input, INPUT_PCD, epsg, 1000)
     assert len(points_clipped) == 678  # check the entity's number of points
+
+
+@pytest.mark.parametrize(
+    "input_file, epsg",
+    # Test on the same geomtries contained in various geometry formats
+    [
+        (INPUT_LIGNES_SHAPE, "EPSG:2154"),  # should work when providing an epsg value + shapefile
+        (INPUT_LIGNES_2D_GEOJSON, "EPSG:2154"),  # should work when providing an epsg value + GeoJSON 2D
+        (INPUT_LIGNES_3D_GEOJSON, "EPSG:2154"),  # should work when providing an epsg value + GeoJSON 3D
+        (INPUT_LIGNES_SHAPE, None),  # Should also work with no epsg value (get from las file) + shapefile
+        (INPUT_LIGNES_2D_GEOJSON, None),  # Should also work with no epsg value (get from las file) + GeoJSON 2D
+        (INPUT_LIGNES_3D_GEOJSON, None),  # Should also work with no epsg value (get from las file) + GeoJSON 3D
+    ],
+)
+def test_clip_3d_lines_to_tile(input_file, epsg):
+    # With lines contained in the LIDAR tile
+    lines_input = gpd.read_file(input_file)
+    lines_clipped = add_points_in_pointcloud.clip_3d_lines_to_tile(lines_input, INPUT_PCD, epsg, 1000)
+    assert len(lines_clipped) == 22  # check the entity's number of lines
+
+    # Without lines contained in the LIDAR tile
+    lines_input = gpd.read_file(input_file)
+    lines_clipped = add_points_in_pointcloud.clip_3d_lines_to_tile(lines_input, INPUT_PCD_CROPPED, epsg, 1000)
+    assert len(lines_clipped) == 0  # check the entity's number of lines
 
 
 @pytest.mark.parametrize(
@@ -56,7 +91,7 @@ def test_add_points_to_las(input_file, epsg, expected_nb_points):
     if Path(OUTPUT_FILE).exists():
         os.remove(OUTPUT_FILE)
 
-    points = gpd.read_file(INPUT_POINTS)
+    points = gpd.read_file(INPUT_POINTS_2D)
     add_points_in_pointcloud.add_points_to_las(points, input_file, OUTPUT_FILE, epsg, 68)
     assert Path(OUTPUT_FILE).exists()  # check output exists
 
@@ -65,40 +100,180 @@ def test_add_points_to_las(input_file, epsg, expected_nb_points):
 
 
 @pytest.mark.parametrize(
-    "input_file, input_points, epsg, expected_nb_points",
+    "line, spacing, z_value, expected_points",
     [
-        (INPUT_PCD, INPUT_POINTS, None, 678),  # should add only points within tile extent
-        (INPUT_PCD_CROPPED, INPUT_POINTS_FOR_CROPPED_PCD, None, 186),
+        # End point is a multiple of spacing, z_value is provided
         (
-            INPUT_PCD_CROPPED,
-            INPUT_POINTS,
-            None,
-            0,
-        ),  # Should add no points when there is only points outside the tile extent
+            LineString([(0, 0), (4, 0)]),
+            2,
+            0.5,
+            [
+                Point(0, 0, 0.5),
+                Point(2, 0, 0.5),
+                Point(4, 0, 0.5),
+            ],
+        ),
+        # End point is not a multiple of spacing, z_value is provided
         (
-            INPUT_PCD_CROPPED,
-            INPUT_POINTS_FOR_CROPPED_PCD,
-            "EPSG:2154",
-            186,
-        ),  # Should work with or without an input epsg
+            LineString([(9, 0), (9, 9)]),
+            5,
+            0.5,
+            [Point(9, 0, 0.5), Point(9, 5, 0.5), Point(9, 9, 0.5)],
+        ),
+        # End point is not a multiple of spacing, z_value is provided in point.z instead of z_value
+        (
+            LineString([(0, 0, 1), (4, 0, 1)]),
+            2,
+            0.5,
+            [
+                Point(0, 0, 1),
+                Point(2, 0, 1),
+                Point(4, 0, 1),
+            ],
+        ),
     ],
 )
-def test_add_points_from_geojson_to_las(input_file, input_points, epsg, expected_nb_points):
+def test_line_to_multipoint(line, spacing, z_value, expected_points):
+    multipoint = add_points_in_pointcloud.line_to_multipoint(line, spacing, z_value)
+    assert multipoint.equals(MultiPoint(expected_points))
+
+
+@pytest.mark.parametrize(
+    "lines_gdf, spacing, altitude_column, expected_points",
+    [
+        # Test case for 2D lines with Z values in "RecupZ"
+        (
+            gpd.GeoDataFrame(
+                {"geometry": [LineString([(0, 0), (10, 0)]), LineString([(10, 0), (10, 10)])], "RecupZ": [5.0, 10.0]},
+                crs="EPSG:2154",
+            ),
+            2.5,
+            "RecupZ",
+            [
+                Point(0, 0, 5.0),
+                Point(2.5, 0, 5.0),
+                Point(5.0, 0, 5.0),
+                Point(7.5, 0, 5.0),
+                Point(10, 0, 5.0),
+                Point(10, 2.5, 10.0),
+                Point(10, 5.0, 10.0),
+                Point(10, 7.5, 10.0),
+                Point(10, 10, 10.0),
+            ],
+        ),
+        # Test case for 3D lines
+        (
+            gpd.GeoDataFrame(
+                {"geometry": [LineString([(0, 0, 3), (10, 0, 3)]), LineString([(10, 0, 6), (10, 10, 6)])]},
+                crs="EPSG:2154",
+            ),
+            2.5,
+            None,
+            [
+                Point(0, 0, 3.0),
+                Point(2.5, 0, 3.0),
+                Point(5.0, 0, 3.0),
+                Point(7.5, 0, 3.0),
+                Point(10, 0, 3.0),
+                Point(10, 2.5, 6.0),
+                Point(10, 5.0, 6.0),
+                Point(10, 7.5, 6.0),
+                Point(10, 10, 6.0),
+            ],
+        ),
+    ],
+)
+def test_generate_3d_points_from_lines(lines_gdf, spacing, altitude_column, expected_points):
+    points_gdf = add_points_in_pointcloud.generate_3d_points_from_lines(lines_gdf, spacing, altitude_column)
+
+    # Check the result
+    assert points_gdf.geometry.tolist() == expected_points
+
+
+@pytest.mark.parametrize(
+    "input_file, input_points, epsg, expected_nb_points, spacing, altitude_column",
+    [
+        (INPUT_PCD, INPUT_POINTS_2D, "EPSG:2154", 678, 0, "RecupZ"),  # should add only points 2.5D within tile extent
+        (INPUT_PCD, INPUT_POINTS_3D, "EPSG:2154", 678, 0, None),  # should add only points 3D within tile extent
+        (INPUT_PCD_CROPPED, INPUT_POINTS_3D_FOR_CROPPED_PCD, "EPSG:2154", 186, 0, None),
+        (INPUT_PCD_CROPPED, INPUT_POINTS_2D_FOR_CROPPED_PCD, "EPSG:2154", 186, 0, "RecupZ"),
+        (
+            INPUT_PCD,
+            INPUT_LIGNES_2D_GEOJSON,
+            "EPSG:2154",
+            678,
+            0.25,
+            "RecupZ",
+        ),  # should add only lines (.GeoJSON) within tile extend
+        (
+            INPUT_PCD,
+            INPUT_LIGNES_SHAPE,
+            "EPSG:2154",
+            678,
+            0.25,
+            "RecupZ",
+        ),  # should add only lignes (.shp) within tile extend
+        (INPUT_PCD, INPUT_LIGNES_SHAPE, None, 678, 0.25, "RecupZ"),  # Should work with or with an input epsg
+        (
+            INPUT_PCD,
+            INPUT_LIGNES_3D_GEOJSON,
+            None,
+            678,
+            0.25,
+            None,
+        ),  # Should work with or without an input epsg and without altitude_column
+    ],
+)
+def test_add_points_from_geometry_to_las(input_file, input_points, epsg, expected_nb_points, spacing, altitude_column):
     # Ensure the output file doesn't exist before the test
     if Path(OUTPUT_FILE).exists():
         os.remove(OUTPUT_FILE)
 
-    add_points_in_pointcloud.add_points_from_geojson_to_las(input_points, input_file, OUTPUT_FILE, 68, epsg, 1000)
+    add_points_in_pointcloud.add_points_from_geometry_to_las(
+        input_points, input_file, OUTPUT_FILE, 68, epsg, 1000, spacing, altitude_column
+    )
     assert Path(OUTPUT_FILE).exists()  # check output exists
     point_count = compute_count_one_file(OUTPUT_FILE)["68"]
     assert point_count == expected_nb_points  # Add all points from geojson
+
+
+@pytest.mark.parametrize(
+    "input_file, input_points, epsg, spacing, altitude_column",
+    [
+        (INPUT_PCD, INPUT_LIGNES_SHAPE, None, 0, "RecupZ"),  # spacing <= 0
+        (INPUT_PCD, INPUT_LIGNES_SHAPE, None, -5, "RecupZ"),  # spacing < 0
+        (
+            INPUT_PCD,
+            INPUT_LIGNES_3D_GEOJSON,
+            None,
+            0,
+            None,
+        ),  # spacing <= 0
+    ],
+)
+def test_add_points_from_geometry_to_las_nok(input_file, input_points, epsg, spacing, altitude_column):
+    # Ensure the output file doesn't exist before the test
+    if Path(OUTPUT_FILE).exists():
+        os.remove(OUTPUT_FILE)
+
+    with pytest.raises(NotImplementedError, match=".*LineString.*spacing.*"):
+        add_points_in_pointcloud.add_points_from_geometry_to_las(
+            input_points,
+            input_file,
+            OUTPUT_FILE,
+            68,
+            epsg,
+            1000,
+            spacing,
+            altitude_column,
+        )
 
 
 def test_parse_args():
     # sanity check for arguments parsing
     args = add_points_in_pointcloud.parse_args(
         [
-            "--input_geojson",
+            "--input_geometry",
             "data/points_3d/Points_virtuels_0292_6833.geojson",
             "--input_las",
             "data/decimated_laz/test_semis_2023_0292_6833_LA93_IGN69.laz",
@@ -110,8 +285,12 @@ def test_parse_args():
             "EPSG:2154",
             "--tile_width",
             "1000",
+            "--spacing",
+            "0",
+            "--altitude_column",
+            "RecupZ",
         ]
     )
     parsed_args_keys = args.__dict__.keys()
-    main_parameters = inspect.signature(add_points_in_pointcloud.add_points_from_geojson_to_las).parameters.keys()
+    main_parameters = inspect.signature(add_points_in_pointcloud.add_points_from_geometry_to_las).parameters.keys()
     assert set(parsed_args_keys) == set(main_parameters)